refactor(trackerForm): rename completions methods file
