+++ src/components/forms/RequiredCompletionsForm/RequiredCompletionsForm.tsx
@@ -23,7 +23,7 @@ import CompletionUnitTextField from '../completions/CompletionUnitTextField';
 import {
   computeDecrementedStringQuantity,
   computeIncrementedStringQuantity
-} from '../completions/computeNewQuantity';
+} from '../completions/utils';
 
 export const FieldsetGrid = styled(Grid)`
   border: 1px solid rgba(0, 0, 0, 0.23);
